requirements.txt
alembic
discord
flake8
gspread
oauth2client
pydantic
pysqlite3
pytest
python-dotenv
sqlalchemy